Add the package manifest file to inform Unity of this package's information.
The package manifest manages information about a version of a package - as one
might expect. It's a bit like `npm`'s manifest files, but also a bit different.
Unity's specifications are recorded at [1], but an inspection of the manifests
in Unity's own packages reveals a few other options, such as "samples" and
"relatedPackages" - it's unclear to what extent these attribute are public or
supported currently or in the future. For now, we'll work with these attributes
but keep the "secret" ones in mind for the future.

The "unity" and "unityRelease" values here really don't mean anything - we have
not tested the example code on any particular versions of Unity.

[1]: https://docs.unity3d.com/Manual/upm-manifestPkg.html