Add sample C# and assembly definitions for the Runtime and Editor environments.
This follows the layout recommended by [1]. It enables the scripts in the
Editor/ folder to have `using UnityEditor;` (it seems because "Editor" is in the
"includePlatforms").

For more on assembly definitions in packages, see [2]. Overall, there are _many_
more configurable features of assembly definitions, some of which we'll almost
certainly want to use. A good overview is at [3], but we'll intentionally keep
it simple until we bundle the actual complex C# SDK project.

[1]: https://docs.unity3d.com/Manual/cus-layout.html
[2]: https://docs.unity3d.com/Manual/cus-asmdef.html
[3]: https://docs.unity3d.com/Manual/ScriptCompilationAssemblyDefinitionFiles.html